+++ src/SFA.DAS.Payments.AcceptanceTests/SFA.DAS.Payments.AcceptanceTests.csproj
@@ -121,7 +121,10 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RowKeys.cs" />
     <Compile Include="SpecFlowHooks.cs" />
+    <Compile Include="StepDefinitions\Base\BaseCalculationSteps.cs" />
     <Compile Include="StepDefinitions\EarningAndPaymentsSteps.cs" />
+    <Compile Include="StepDefinitions\EarningStepDefinitions.cs" />
+    <Compile Include="StepDefinitions\PaymentStepDefinitions.cs" />
     <Compile Include="StepDefinitions\ReferenceDataSteps.cs" />
     <Compile Include="ExecutionEnvironment\TestLogger.cs" />
     <Compile Include="Translators\IlrTranslator.cs" />
